Move serialized types to their own crate
For #117
