+++ cargo-geiger-serde/Cargo.toml
@@ -6,6 +6,7 @@ license = "Apache-2.0/MIT"
 name = "cargo-geiger-serde"
 version = "0.1.0"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
+semver = "0.10.0"
+serde = { version = "1.0.116", features = ["derive"] }
+url = { version = "2.1.1", features = ["serde"] }
